ignore SIGPIPE
On termination, when Boulevard is supervised by another process (e.g. with
daemon(8) on FreeBSD), the parent will close file descriptors such as
stdout/stderr without waiting for Boulevard to exit. Boulevard logs messages to
stderr during shutdown, leading to SIGPIPE. SIGPIPE causes the process to be
killed instantly. For most programs is does not matter, but Boulevard must be
able to close the control API UNIX socket or it will remain present in the
filesystem and prevent any restart.

Feel free to blame daemon(8), the ridiculous default effect of SIGPIPE, or the
idiocy of UNIX sockets not being automatically deleted on process termination.
Yes UNIX sucks.